requirements.txt
flask
google-generativeai
openai
python-dotenv
gunicorn
flask-socketio
python-telegram-bot
